test(admin): move ProjectStep tests to src/tests/components/
Match the admin package's dominant test convention. ~30+ existing
component tests live in `src/tests/components/`; only a handful of
outliers (rule-builder, utils/*) stayed colocated. New ProjectStep
tests were placed next to the component by following one of the
outliers — moving them to the central location for consistency.

Updated imports: `./project-step` → `../../components/integrations/
project-step`. `vi.mock` paths (`../../services/...`, `../../i18n/
...`) resolve to the same targets from the new depth.

751/751 admin unit tests pass; lint + build clean.
